fix: bug修复

+++ src/module/category/category.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
 import { CategoryModel } from './category.model'
+import { ErrorException } from '@/filter/error-exception.filter'
+import { ErrorCode } from '@/filter/error-code'
 
 @Injectable()
 export class CategoryService {
@@ -9,6 +11,27 @@ export class CategoryService {
     private categoryModel: typeof CategoryModel
   ) {}
   async create(data) {
+    const { parent_id = 0, name } = data
+    if (parent_id !== 0) {
+      const parent = await this.categoryModel.findOne({
+        where: {
+          id: parent_id
+        }
+      })
+      if (!parent) {
+        throw new ErrorException(ErrorCode[991003])
+      }
+    }
+
+    const findOne = await this.categoryModel.findOne({
+      where: {
+        name
+      }
+    })
+    if (findOne) {
+      throw new ErrorException(ErrorCode[991002])
+    }
+
     return this.categoryModel.create(data)
   }
 
@@ -37,15 +60,32 @@ export class CategoryService {
 
   async update(data) {
     const { id, ..._data } = data
+
+    const findOne = await this.findOne(data)
+    if (!findOne) {
+      throw new ErrorException(ErrorCode[991001])
+    }
     return this.categoryModel.update(_data, { where: { id } })
   }
 
   async delete(data) {
     const { id } = data
+
+    const findOne = await this.categoryModel.findOne({
+      where: {
+        id
+      },
+      paranoid: false
+    })
+    if (!findOne) {
+      throw new ErrorException(ErrorCode[991001])
+    }
+
     return this.categoryModel.destroy({
       where: {
         id
-      }
+      },
+      force: true
     })
   }
 }
